chore(settlement): move inline test modules to sibling files
Move four inline #[cfg(test)] mod tests blocks into sibling tests.rs
files so upcoming PRs in this stack have clean diffs. Pure move, zero
semantic changes: bodies de-indented one level plus rustfmt reflow of
lines that gained width. cargo nextest list is byte-identical before
and after (348 tests).

Preparation for splitting #1663 into incremental PRs. Part of #1675.
